modules/IR/requirements.txt
certifi==2025.11.12
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@19870eafbf315e6bfbc73fc97935abbd83924416
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
faiss-cpu==1.13.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
netCDF4==1.7.3
networkx==3.6
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pyparsing==3.2.5
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
textblob==0.19.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
